it is uint32_t, so leave it like that

+++ tt_metal/hw/inc/api/tensor/tensor_accessor.h
@@ -278,7 +278,7 @@ private:
     uint16_t get_bank_y(uint16_t packed_xy_coord) const { return packed_xy_coord & 0xFF; }
 
 public:
-    const uintptr_t bank_base_address = 0;
+    const uint32_t bank_base_address = 0;
     const uint32_t page_size = 0;
 
     friend class tensor_accessor::ShardPagesAddressIterator<TensorAccessor>;
@@ -314,7 +314,7 @@ struct TensorAccessor<tensor_accessor::DistributionSpec<
     template <std::size_t CTA_OFFSET, std::size_t CRTA_OFFSET>
     TensorAccessor(
         const TensorAccessorArgs<CTA_OFFSET, CRTA_OFFSET>& args,
-        const uintptr_t bank_base_address_in,
+        const uint32_t bank_base_address_in,
         const uint32_t page_size_in = 0) :
         InterleavedAddrGen<IsDram>(
             {.bank_base_address = static_cast<uint32_t>(bank_base_address_in), .page_size = page_size_in}) {}
